feat: Add Pascal program for saving screen and related unit
This commit adds a new Pascal program called "saveScreen" that allows users to save the screen contents. It also includes a related unit called "unit_saveScreen" that provides functions for writing the title and protecting the screen.

The "saveScreen" program uses the "unit_saveScreen" unit to write a title and protect the screen. It then waits for user input before exiting.

This commit addresses the need for a screen-saving functionality in Pascal programs.